rework charts/dcrpg/online test build tags
